Rename GhostPingHandler to GhostPingDeleteHandler and update setFooter in embed to remove deprecation warning
